% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_frontpage.R
\name{get_frontpage}
\alias{get_frontpage}
\title{Get links in news section of morningconsult.com}
\usage{
get_frontpage()
}
\value{
list of tibbles - main page story links and links to page sections
}
\description{
Story links from \code{news} section
}
\examples{
\dontrun{
links <- get_frontpage()
}
}
