% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links.R
\name{get_links}
\alias{get_links}
\title{Get all links from morningconsult.com}
\usage{
get_links()
}
\value{
tibble of all story links
}
\description{
Get all links from morningconsult.com
}
\examples{
\dontrun{
links <- get_links()
}
}
