% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sectionstories.R
\name{get_sectionstories}
\alias{get_sectionstories}
\title{Get links from morningconsult sections}
\usage{
get_sectionstories(...)
}
\arguments{
\item{ll}{}
}
\value{
tibble of links
}
\description{
Get links from morningconsult sections
}
\examples{
\dontrun{
link <- "morningconsult.com/coronavirus"
ss_links <- get_sectionstories(link)
}
}
